Subject: [rt.cpan.org #44886] ::Property::_fold does not follow RFC Date: Wed, 08 Apr 2009 21:30:49 -0400 From: Jerrad Pierce via RT <bug-Data-ICal@rt.cpan.org> To: undisclosed-recipients: ; Reply-To: bug-Data-ICal@rt.cpan.org RT-Originator: jpierce@cpan.org
Wed Apr 08 21:30:48 2009: Request 44886 was acted upon.
Transaction: Ticket created by JPIERCE
       Queue: Data-ICal
     Subject: ::Property::_fold does not follow RFC
   Broken in: 0.13, 0.15
    Severity: Important
       Owner: Nobody
  Requestors: jpierce@cpan.org
      Status: new
 Ticket <URL: https://rt.cpan.org/Ticket/Display.html?id=44886 >

The RFC says to fold with CRLF, not LF.
This can compound problems with some parsers.

Below is a patch (against .13) to correct this. It also
uses a tab instead of a space so that folded lines are
more human readable, as well as trying to not break words.